@@ -12,14 +12,16 @@
     "test:example": "npm run build && playwright test tests"
   },
   "dependencies": {
-    "ipfs": "^0.57.0",
-    "next": "11.0.1",
+    "ipfs": "^0.58.1",
+    "next": "11.0.0",
     "react": "17.0.2",
     "react-dom": "17.0.2"
   },
   "devDependencies": {
-    "eslint": "7.30.0",
-    "eslint-config-next": "11.0.1",
+    "@playwright/test": "^1.12.3",
+    "eslint": "7.32.0",
+    "eslint-config-next": "11.0.0",
+    "playwright": "^1.12.3",
     "rimraf": "^3.0.2",
     "test-util-ipfs-example": "^1.0.0"
   }
