@@ -15,12 +15,14 @@
     "clean": "echo 'Nothing to clean...'",
     "start": "electron .",
     "serve": "npm run start",
-    "test:example": "node tests/test.js"
+    "test:example": "node test.js"
+  },
+  "dependencies": {
+    "ipfs": "^0.58.1"
   },
   "devDependencies": {
-    "electron": "^12.0.2",
-    "electron-rebuild": "^2.3.4",
-    "ipfs": "^0.57.0",
+    "electron": "^13.1.9",
+    "electron-rebuild": "^3.1.1",
     "test-util-ipfs-example": "^1.0.2"
   },
   "greenkeeper": {
