@@ -5,13 +5,14 @@
   "description": "Template for using js-ipfs in typescript project",
   "scripts": {
     "clean": "echo 'Nothing to clean...'",
-    "test": "tsc --noEmit"
+    "test": "tsc --noEmit",
+    "test:example": "npm run test"
   },
   "dependencies": {
-    "ipfs": "^0.57.0",
+    "ipfs": "^0.58.1",
     "multiformats": "^9.4.1"
   },
   "devDependencies": {
-    "typescript": "4.2.x"
+    "typescript": "4.3.5"
   }
 }
