@@ -213,7 +213,7 @@ function appendFile (name, hash, size, data) {
   const link = document.createElement('a')
   link.setAttribute('href', url)
   link.setAttribute('download', name)
-  link.innerHTML = '<img width=20 class="table-action" src="assets/download.svg" alt="Download" />'
+  link.innerHTML = '<img width=20 class="table-action" src="download.svg" alt="Download" />'
   downloadCell.appendChild(link)
 
   row.appendChild(nameCell)
