@@ -2,7 +2,7 @@
 
 const webRTCStarSigServer = require('libp2p-webrtc-star/src/sig-server')
 
-const { test } = require('@playwright/test');
+const { test, expect } = require('@playwright/test');
 const { playwright } = require('test-util-ipfs-example');
 
 // Setup
@@ -45,6 +45,8 @@ const play = test.extend({
   )
 });
 
+test.setTimeout(1000 * 60 * 2);
+
 play.describe('upload file using http client: ', () => {
   // DOM
   const nodeId = '.node-id'
@@ -64,7 +66,7 @@ play.describe('upload file using http client: ', () => {
   const FILE_CONTENT = 'A file with some content'
   const WORKSPACE = 'test-workspace'
 
-  const connect = async (page, address, workspace, remotePeerId) => {
+  const connect = async (page, address) => {
     await page.waitForSelector(connectedPeers)
 
     await page.fill(multiAddressInput, address)
@@ -72,11 +74,11 @@ play.describe('upload file using http client: ', () => {
 
     const IPFS_RELAY_ID = address.split('/').pop()
     await page.waitForSelector(`${peers}:has-text('${IPFS_RELAY_ID}')`)
+  }
 
+  const subscribe = async (page, workspace) => {
     await page.fill(workspaceInput, workspace)
     await page.click(workspaceBtn)
-
-    await page.waitForSelector(`${workspacePeers}:has-text('${remotePeerId}')`)
   }
 
   let sigServer;
@@ -116,12 +118,25 @@ play.describe('upload file using http client: ', () => {
       .map(ma => ma.toString())
       .find(addr => addr.includes('/ws/p2p/'))
 
+    if (!IPFS_RELAY_ADDRESS) {
+      throw new Error(`Could not find web socket address in ${id.addresses}`)
+    }
 
     const pageOnePeerId = (await pageOne.textContent(nodeId)).trim()
     const pageTwoPeerId = (await pageTwo.textContent(nodeId)).trim()
 
-    await connect(pageOne, IPFS_RELAY_ADDRESS, WORKSPACE, pageTwoPeerId)
-    await connect(pageTwo, IPFS_RELAY_ADDRESS, WORKSPACE, pageOnePeerId)
+    await connect(pageOne, IPFS_RELAY_ADDRESS)
+    await connect(pageTwo, IPFS_RELAY_ADDRESS)
+
+    await subscribe(pageOne, WORKSPACE, pageTwoPeerId)
+    await subscribe(pageTwo, WORKSPACE, pageOnePeerId)
+
+    await pageOne.waitForSelector(`${workspacePeers}:has-text('${pageTwoPeerId}')`, {
+      timeout: 10000
+    })
+    await pageTwo.waitForSelector(`${workspacePeers}:has-text('${pageOnePeerId}')` , {
+      timeout: 10000
+    })
 
     // only one browser should add the file to the workspace
     await pageOne.fill(cidInput, cid.toString())
@@ -129,7 +144,21 @@ play.describe('upload file using http client: ', () => {
 
     // but should both see the added file
     await pageOne.waitForSelector(`${fileHistory}:has-text('${cid.toString()}')`)
-    await pageTwo.waitForSelector(`${fileHistory}:has-text('${cid.toString()}')`)
+    let passed = false;
+
+    for (let index = 0; index < 3; index++) {
+      await pageTwo.waitForTimeout(15000)
+      const contentHistory = await pageTwo.textContent(fileHistory)
+      passed = contentHistory.includes(cid.toString())
+
+      if (passed) {
+        break
+      } else {
+        await pageTwo.reload();
+      }
+    }
+
+    expect(passed).toBeTruthy()
   });
 
 });
