@@ -12,13 +12,13 @@ const play = test.extend({
 
 play.describe('browser videostream:', () => {
   // DOM
-  const video = "#video"
+  const status = "#status"
 
   play.beforeEach(async ({servers, page}) => {
     await page.goto(`http://localhost:${servers[0].port}/`);
   })
 
   play('should properly initialized a IPFS node and stream a video', async ({ page }) => {
-    await page.waitForFunction((el) => document.querySelector(el).readyState >= 2, video);
+    await page.waitForSelector(`${status}:has-text('Video ready')`)
   });
 });
