@@ -5,7 +5,7 @@ const { node } = require('test-util-ipfs-example');
 const Libp2p = require('libp2p')
 const TCP = require('libp2p-tcp')
 const MPLEX = require('libp2p-mplex')
-const { NOISE } = require('libp2p-noise')
+const { NOISE } = require('@chainsafe/libp2p-noise')
 const PeerId = require('peer-id')
 const uint8ArrayToString = require('uint8arrays/to-string')
 
