@@ -1,51 +1,55 @@
-'use strict'
+"use strict";
 
-const { app, BrowserWindow } = require('electron')
-const IPFS = require('ipfs')
+const { app, BrowserWindow } = require("electron");
+const IPFS = require("ipfs");
 
-let mainWindow
+let mainWindow;
 
-function createWindow () {
+function createWindow() {
   mainWindow = new BrowserWindow({
     width: 800,
     height: 600,
     webPreferences: {
       nodeIntegration: true
     }
-  })
+  });
 
   // and load the index.html of the app.
-  mainWindow.loadFile('index.html')
+  mainWindow.loadFile("index.html");
 
   // Open the DevTools.
-  mainWindow.webContents.openDevTools()
+  mainWindow.webContents.openDevTools();
 
   // Emitted when the window is closed.
-  mainWindow.on('closed', () => {
-    mainWindow = null
-  })
+  mainWindow.on("closed", () => {
+    mainWindow = null;
+  });
 }
 
-app.on('ready', async () => {
-  createWindow()
+app.on("ready", async () => {
+  createWindow();
 
   try {
-    const node = await IPFS.create()
-    const id = await node.id()
-    console.log(id)
+    const node = await IPFS.create();
+    const id = await node.id();
+    console.log(id);
   } catch (err) {
-    console.error(err)
+    console.error(err);
   }
-})
+});
 
 // Quit when all windows are closed.
-app.on('window-all-closed', () => {
-  if (process.platform !== 'darwin') { app.quit() }
-})
+app.on("window-all-closed", () => {
+  if (process.platform !== "darwin") {
+    app.quit();
+  }
+});
 
-app.on('activate', () => {
-  if (mainWindow === null) { createWindow() }
-})
+app.on("activate", () => {
+  if (mainWindow === null) {
+    createWindow();
+  }
+});
 
 // In this file you can include the rest of your app's specific main process
 // code. You can also put them in separate files and require them here.
